tests building and running

+++ src/chachapoly.c
@@ -99,7 +99,7 @@ static void chacha20_poly1305_aead_encrypt(struct ssh_cipher_struct *cipher,
     chacha_encrypt_bytes(&keys->k2,
                          in_packet->payload,
                          out_packet->payload,
-                         len - sizeof(uint32_t));
+                         (uint32_t)(len - sizeof(uint32_t)));
 
     /* ssh_log_hexdump("poly1305_ctx", poly1305_ctx, sizeof(poly1305_ctx)); */
     /* step 4, compute the MAC */
@@ -174,7 +174,7 @@ static int chacha20_poly1305_aead_decrypt(struct ssh_cipher_struct *cipher,
     chacha_encrypt_bytes(&keys->k2,
                          (uint8_t *)complete_packet + sizeof(uint32_t),
                          out,
-                         encrypted_size);
+                         (uint32_t)encrypted_size);
 
     return SSH_OK;
 }
